// src/api/generated/types.gen.ts
export type OrderStatus =
  'PENDING' | 'PAID' | 'SHIPPED' | 'DELIVERED' | 'CANCELLED' | 'REFUNDED';

export type OrderItemResponseDto = {
  id: string;
  orderId: string;
  productId: string;
  name: string;
  priceCents: number;
  quantity: number;
};

export type OrderResponseDto = {
  status: OrderStatus;
  stripeSessionId: string | null;
  stripePaymentIntentId: string | null;
  shippingAddress: {
    [key: string]: unknown;
  } | null;
  totalCents: number | null;
  items: Array<OrderItemResponseDto>;
  id: string;
  userId: string;
  email: string;
  subtotalCents: number;
  createdAt: string;
  updatedAt: string;
};